Fix coding standard violations
