+++ src/main/java/dukebot/MainWindow.java
@@ -1,4 +1,4 @@
-package DukeBot;
+package dukebot;
 
 import javafx.application.Platform;
 import javafx.fxml.FXML;
