+++ src/main/java/dukebot/TaskList.java
@@ -1,4 +1,4 @@
-package DukeBot;
+package dukebot;
 
 import java.util.ArrayList;
