+++ src/main/java/dukebot/ToDo.java
@@ -1,4 +1,4 @@
-package DukeBot;
+package dukebot;
 
 /**
  * Encapsulates the ToDo subclass of Task.
