+++ src/main/java/dukebot/command/ListCommand.java
@@ -1,8 +1,8 @@
-package DukeBot.command;
+package dukebot.command;
 
-import DukeBot.DukeException;
-import DukeBot.TaskList;
-import DukeBot.Ui;
+import dukebot.DukeException;
+import dukebot.TaskList;
+import dukebot.Ui;
 
 public class ListCommand extends Command {
